Steam Key Reader
This is the beginning of a program that when ran, willl scan websites for Steam game keys. The Steam API has been imported as well. The program will be able to scan webpages and submit them to steam on it's own with your own log in information.

I'm currently working on getting regex to work via reddit's homepage.